Move hw_breakpoint to its own file
